Removed Option<ch>.
It's only used in the ansi drawer. There is a new invisible char to represent that case.
That is 1:1 the dos default char with another attr set.

So old code that uses .unwrap_or_default() w/o additional checks is still working.

This makes the library just easier to use every speed test I did shows that there is no difference.
Rust is very fast with it's Option type :).